+++ Public/Invoke-PSDuckDb.ps1
@@ -103,5 +103,4 @@ USING sum(Population);
             }
         }
     }
-
 }