+++ changelog.md
@@ -19,6 +19,7 @@
 - add ordered to hashtable to keep order of column names
 - dispose of reader and duckCommand
 
+
 # 0.1.0
 
 - Initial release